+++ src/main/java/com/jdgorman/fooddeliveryapi/enumerator/CustomerOrderStatus.java
@@ -1,0 +1,11 @@
+package com.jdgorman.fooddeliveryapi.enumerator;
+
+public enum CustomerOrderStatus {
+    PENDING,
+    CONFIRMED,
+    PREPARING,
+    READY,
+    OUT_FOR_DELIVERY,
+    DELIVERED,
+    CANCELLED
+}