+++ src/test/java/com/jdgorman/fooddeliveryapi/service/MenuItemServiceTest.java
@@ -2,7 +2,7 @@ package com.jdgorman.fooddeliveryapi.service;
 
 import com.jdgorman.fooddeliveryapi.dto.MenuItemRequest;
 import com.jdgorman.fooddeliveryapi.dto.MenuItemResponse;
-import com.jdgorman.fooddeliveryapi.entity.MenuCategory;
+import com.jdgorman.fooddeliveryapi.enumerator.MenuCategory;
 import com.jdgorman.fooddeliveryapi.entity.MenuItem;
 import com.jdgorman.fooddeliveryapi.entity.Restaurant;
 import com.jdgorman.fooddeliveryapi.exception.ResourceNotFoundException;